Add files via upload
We created two dashboards for our university library at WPI. These dashboards monitor the library's occupancy and track the expenditure on student employees. Tableau was used to design and develop these dashboards. The dashboard analyses which parts are most used or beneficial, we have also done a conjoint analysis to understand which zones and seats do the students prefer.